Enable transparency drawing
Objects are drawn in two steps, by Object::draw() and
Object::drawLast(). This allow transparent surfaces to be drawn after
opaque surfaces.

This is not perfect since several layers of transparent surfaces may be
rendered differently depending on object order. But it should be enough
for most use cases.

OSimple uses the appropriated method depending on its color.